Bug Fix: BLAST broken pipe
Shouldn’t have trusted the Shell to redirect output. The redirect >>
now goes through GNU parallel, which should be able to handle it when
the results collide. Bug only appeared when there were tons of hits.
The -k setting in GNU parallel ensures that the jobs print in the order
they are called, which should mean the results file would not be
written two by two jobs simultaneously.